extract brand banner out.
