+++ apps/vor/frontend/src/components/BrandBanner/BrandBanner.stories.tsx
@@ -1,0 +1,12 @@
+import React, { FC } from "react"
+import BrandBanner from "./BrandBanner"
+
+export default {
+  title: "Core|BrandBanner",
+  component: BrandBanner,
+  parameters: {
+    componentSubtitle: "Just a simple BrandBanner",
+  },
+}
+
+export const Basic: FC = () => <BrandBanner />
