+++ apps/vor/frontend/src/components/BrandBanner/BrandBanner.test.tsx
@@ -1,0 +1,10 @@
+import React from "react"
+import { render } from "../../test-utils"
+import { Basic } from "./BrandBanner.stories"
+
+describe("BrandBanner", () => {
+  it("should render correctly", () => {
+    const { container } = render(<Basic />)
+    expect(container).toMatchSnapshot()
+  })
+})
