+++ apps/vor/frontend/src/components/PageLogin/PageLogin.tsx
@@ -1,13 +1,13 @@
 /** @jsx jsx */
-import { FC, FormEvent, useState, Fragment } from "react"
-import { graphql, navigate, useStaticQuery } from "gatsby"
+import { FC, FormEvent, useState } from "react"
+import { navigate } from "gatsby"
 import { Box, Button, Flex, jsx } from "theme-ui"
-import GatsbyImage from "gatsby-image"
 import Input from "../Input/Input"
 import { Link } from "../Link/Link"
 import { Typography } from "../Typography/Typography"
 import Icon from "../Icon/Icon"
 import { ReactComponent as InfoIcon } from "../../icons/info.svg"
+import BrandBanner from "../BrandBanner/BrandBanner"
 
 export const PageLogin: FC = () => {
   const [email, setEmail] = useState("")
@@ -40,7 +40,7 @@ export const PageLogin: FC = () => {
 
   return (
     <Box>
-      <Header />
+      <BrandBanner />
 
       <Flex sx={{ justifyContent: "center" }}>
         <Box
@@ -126,51 +126,6 @@ export const PageLogin: FC = () => {
         </Flex>
       </Box>
     </Box>
-  )
-}
-
-const Header = () => {
-  const query = useStaticQuery(graphql`
-    query {
-      file(relativePath: { eq: "logo-transparent.png" }) {
-        childImageSharp {
-          # Specify the image processing specifications right in the query.
-          # Makes it trivial to update as your page's design changes.
-          fixed(width: 40, height: 40) {
-            ...GatsbyImageSharpFixed
-          }
-        }
-      }
-    }
-  `)
-
-  return (
-    <Fragment>
-      <Flex
-        mx="auto"
-        py={3}
-        px={[3, 0]}
-        sx={{
-          width: "100%",
-          maxWidth: "maxWidth.xsm",
-          alignItems: "center",
-        }}
-      >
-        <GatsbyImage
-          fixed={query.file.childImageSharp.fixed}
-          sx={{ flexShrink: 0 }}
-        />
-        <Typography.Body
-          ml={2}
-          sx={{ fontSize: 3, fontWeight: "bold", lineHeight: 1.2 }}
-        >
-          Obserfy{" "}
-          <span sx={{ fontWeight: "normal", whiteSpace: "nowrap" }}>
-            for Teachers
-          </span>
-        </Typography.Body>
-      </Flex>
-    </Fragment>
   )
 }
